requirements.txt
backports.csv==1.0.7
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.7
django-ckeditor==5.6.1
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-redis==4.10.0
django-reversion==3.0.3
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
future==0.17.1
httplib2==0.9.2
idna==2.8
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
Pillow==5.4.1
pycryptodomex==3.7.2
PyJWT==1.7.1
PyMySQL==0.9.3
python-alipay-sdk==1.10.0
pytz==2018.9
PyYAML==5.1
redis==3.2.1
requests==2.21.0
six==1.12.0
tablib==0.13.0
urllib3==1.24.1
xadmin==2.0.1
xlrd==1.2.0
xlwt==1.3.0
